requirements.txt
# Core dependencies
Django>=5.2.1
djangorestframework>=3.16.0
psycopg2-binary>=2.9.10\

# Swagger
drf-yasg>=1.21.10

# Environment variable management
django-environ>=0.12.0

# Faker
Faker>=37.3.0

# Filtering support for DRF
django-filter>=25.1

# JWT authentication
djangorestframework-simplejwt>=5.5.0

